[MIGRATIONS] make handful of migrations more reliable

+++ migrations/Migration20131106150723ComProjects.php
@@ -15,10 +15,9 @@ class Migration20131106150723ComProjects extends Base
 	 **/
 	public function up()
 	{
-		$query = "ALTER TABLE `#__projects` ADD FULLTEXT KEY `idx_fulltxt_alias_title_about` (`alias`, `title`, `about`);";
-
-		if (!empty($query))
+		if ($this->db->tableExists('#__projects') && !$this->db->tableHasKey('#__projects', 'idx_fulltxt_alias_title_about'))
 		{
+			$query = "ALTER TABLE `#__projects` ADD FULLTEXT KEY `idx_fulltxt_alias_title_about` (`alias`, `title`, `about`);";
 			$this->db->setQuery($query);
 			$this->db->query();
 		}
@@ -29,10 +28,9 @@ class Migration20131106150723ComProjects extends Base
 	 **/
 	public function down()
 	{
-		$query = "ALTER TABLE `#__projects` DROP INDEX `idx_fulltxt_alias_title_about`;";
-
-		if (!empty($query))
+		if ($this->db->tableExists('#__projects') && $this->db->tableHasKey('#__projects', 'idx_fulltxt_alias_title_about'))
 		{
+			$query = "ALTER TABLE `#__projects` DROP INDEX `idx_fulltxt_alias_title_about`;";
 			$this->db->setQuery($query);
 			$this->db->query();
 		}
